requirements.txt
gradio
requests
langchain_community
langgraph
langchain_google_genai
langchain_groq
langchain_huggingface
supabase
websockets
sentence-transformers
wikipedia